pyproject.toml
[tool.poetry]
name = "jibrel-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"

django = "2.2.2"
djangorestframework = "^3.9"
psycopg2-binary = "^2.8"
gunicorn = "^19.9"
zxcvbn = "^4.4"
requests = "^2.22"
celery = "^4.3"
django-celery-results = "^1.1"
phonenumbers = "^8.10"
django-anymail = "^6.1"
pyyaml = "^5.1"
python-magic = "^0.4.15"
boto3 = "^1.9"
django-storages = "^1.7"
onfido = "^4.1"
pycountry = "^19.7"
sentry-sdk = "^0.10.2"
django-cors-headers = "^3.0"
python-decouple = "^3.1"
redis = "^3.3"
krakenex = "^2.1"
dataclasses-json = "^0.3.5"
python-json-logger = "^0.1.11"
django-constance = {version = "^2.4", extras = ["redis"]}
whitenoise = "^4.1"
django-object-actions = "^1.1"
django-nested-admin = "^3.2"
Pillow = "^6.2"
django-admin-tools = "^0.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylama = "^7.7"
mypy = "^0.711.0"
pytest-django = "^3.5"
pytest-cov = "^2.7"
pytest-mock = "^1.10"
factory_boy = "^2.12"
openapi-core = "^0.12.0"
hypothesis = "^4.41"
isort = "^4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
